refactor: eliminate 99% code duplication with shared buildLockEffect method
## Summary
- Extracted shared lock acquisition logic into Effect+LockmanInternal.buildLockEffect
- Eliminated 99% code duplication between lockAndExecute and buildLockedEffect
- Improved naming consistency: buildLockedEffect → buildEffect, withLock → buildLockEffect

## Changes
- **New**: Sources/Lockman/Composable/Effect+LockmanInternal.swift
  - Added buildLockEffect() method with effect builder pattern
  - Moved acquireLock() and handleError() helper methods
- **Modified**: Sources/Lockman/Composable/Effect+Lockman.swift
  - Simplified buildEffect() to call shared buildLockEffect()
  - Reduced from ~280 lines to ~66 lines (76% reduction)
- **Modified**: Sources/Lockman/Composable/LockmanDynamicConditionReducer.swift
  - Updated to use shared Effect.buildLockEffect() method
  - Removed private lockAndExecute() method (~80 lines removed)

## Benefits
- **Maintainability**: Single source of truth for lock acquisition logic
- **Consistency**: Unified naming convention (build* methods)
- **Testability**: Shared logic can be tested independently
- **Architecture**: Clear separation between public API and internal implementation
